Update event tracking for 2019
