+++ HIAPI/Services/EventService.swift
@@ -14,9 +14,9 @@ import Foundation
 import APIManager
 import CoreData
 
-public final class EventService: BaseService {
+public class EventService: BaseService {
 
-    public override static var baseURL: String {
+    public override class var baseURL: String {
         return super.baseURL + "event/"
     }
 
@@ -46,4 +46,25 @@ public final class EventService: BaseService {
         return APIRequest<Favorite>(service: self, endpoint: "favorite/remove/", body: body, method: .POST)
     }
 
+}
+
+public final class TrackingService: EventService {
+
+    public override static var baseURL: String {
+        return super.baseURL + "track/"
+    }
+
+    public static func track(name: String, id: String) -> APIRequest<SimpleRequest> {
+        var body = HTTPBody()
+        body["eventName"] = name
+        body["userId"] = id
+        return APIRequest<SimpleRequest>(service: self, endpoint: "", body: body, method: .POST)
+    }
+
+    internal override static func hiValidate(statusCode: Int) throws {
+        if ![200, 422, 500].contains(statusCode) {
+            let description = HTTPURLResponse.localizedString(forStatusCode: statusCode)
+            throw APIRequestError.invalidHTTPReponse(code: statusCode, description: description)
+        }
+    }
 }
