+++ HackIllinois/FlowControllers/HIApplicationStateController.swift
@@ -75,11 +75,6 @@ extension HIApplicationStateController {
         viewControllers.append(HIScheduleViewController())
         viewControllers.append(HIAnnouncementsViewController())
         viewControllers.append(HIUserDetailViewController())
-
-        if !user.roles.intersection([.staff, .admin]).isEmpty {
-            viewControllers.append(HIScannerViewController())
-        }
-
         return viewControllers
     }
