+++ HackIllinois/ViewControllers/HIBaseScannerViewController.swift
@@ -1,0 +1,136 @@
+//
+//  HIBaseScannerViewController.swift
+//  HackIllinois
+//
+//  Created by HackIllinois Team on 11/26/17.
+//  Copyright © 2017 HackIllinois. All rights reserved.
+//  This file is part of the Hackillinois iOS App.
+//  The Hackillinois iOS App is open source software, released under the University of
+//  Illinois/NCSA Open Source License. You should have received a copy of
+//  this license in a file with the distribution.
+//
+
+import Foundation
+import UIKit
+import AVKit
+
+class HIBaseScannerViewController: HIBaseViewController {
+    var captureSession: AVCaptureSession?
+    var previewLayer: AVCaptureVideoPreviewLayer?
+    let hapticGenerator = UINotificationFeedbackGenerator()
+
+    var loadFailed = false
+    var respondingToQRCodeFound = true
+
+    var adminEventViewController = HIAdminEventViewController()
+}
+
+// MARK: - UIViewController
+extension HIBaseScannerViewController {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupCaptureSession()
+    }
+
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        if loadFailed {
+            presentErrorController(
+                title: "Scanning not supported",
+                message: "Your device does not support scanning a code from an item. Please use a device with a camera.",
+                dismissParentOnCompletion: false
+            )
+        } else if captureSession?.isRunning == false {
+            previewLayer?.frame = view.layer.bounds
+            DispatchQueue.main.async { [weak self] in
+                self?.captureSession?.startRunning()
+            }
+        }
+    }
+
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+    }
+
+    override func viewDidDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        if captureSession?.isRunning == true {
+            DispatchQueue.main.async { [weak self] in
+                self?.captureSession?.stopRunning()
+            }
+        }
+    }
+
+    override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
+        coordinator.animate(alongsideTransition: { [weak self] (_) in
+            self?.setFrameForPreviewLayer()
+        }, completion: nil)
+    }
+}
+
+// MARK: - UINavigationItem Setup
+extension HIBaseScannerViewController {
+    @objc dynamic override func setupNavigationItem() {
+        super.setupNavigationItem()
+        title = "SCANNER"
+    }
+}
+
+// MARK: AVCaptureMetadataOutputObjectsDelegate
+extension HIBaseScannerViewController: AVCaptureMetadataOutputObjectsDelegate {
+    func setupCaptureSession() {
+        captureSession = AVCaptureSession()
+        let metadataOutput = AVCaptureMetadataOutput()
+
+        guard
+            let captureSession = captureSession,
+            let videoCaptureDevice = AVCaptureDevice.default(for: .video),
+            let videoInput = try? AVCaptureDeviceInput(device: videoCaptureDevice),
+            captureSession.canAddInput(videoInput),
+            captureSession.canAddOutput(metadataOutput) else {
+                loadFailed = true
+                return
+        }
+
+        captureSession.addInput(videoInput)
+        captureSession.addOutput(metadataOutput)
+        metadataOutput.setMetadataObjectsDelegate(self, queue: DispatchQueue.main)
+        metadataOutput.metadataObjectTypes = [.qr]
+
+        let previewLayer = AVCaptureVideoPreviewLayer(session: captureSession)
+        previewLayer.videoGravity = .resizeAspectFill
+        self.previewLayer = previewLayer
+        setFrameForPreviewLayer()
+        view.layer.addSublayer(previewLayer)
+    }
+
+    func setFrameForPreviewLayer() {
+        guard let previewLayer = previewLayer else { return }
+
+        previewLayer.frame = view.layer.bounds
+
+        guard previewLayer.connection?.isVideoOrientationSupported == true else { return }
+
+        switch UIApplication.shared.statusBarOrientation {
+        case .portrait, .unknown:
+            previewLayer.connection?.videoOrientation = .portrait
+        case .portraitUpsideDown:
+            previewLayer.connection?.videoOrientation = .portraitUpsideDown
+        case .landscapeLeft:
+            previewLayer.connection?.videoOrientation = .landscapeLeft
+        case .landscapeRight:
+            previewLayer.connection?.videoOrientation = .landscapeRight
+        }
+    }
+
+    func metadataOutput(_ output: AVCaptureMetadataOutput, didOutput metadataObjects: [AVMetadataObject], from connection: AVCaptureConnection) {
+        guard respondingToQRCodeFound else { return }
+        if let code = (metadataObjects.first as? AVMetadataMachineReadableCodeObject)?.stringValue {
+            found(code: code)
+        }
+    }
+
+    @objc func found(code: String) {
+        fatalError("found(code:) must be implemented in a subclass of HIBaseScannerViewController.")
+    }
+}
